refactor: migrated all tests from jest to vitest
