+++ src/__tests__/lib/is-array.test.ts
