+++ src/__tests__/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "moduleResolution": "nodenext",
     "target": "es2018",
     "lib": ["es2015", "dom"]
   }
